task module refactor:separate resource state from job entity
