+++ server/odc-service/src/main/java/com/oceanbase/odc/service/task/resource/K8sPodResource.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.oceanbase.odc.service.resource.k8s;
+package com.oceanbase.odc.service.task.resource;
 
 import java.util.Date;
